refactor(core): move shared node base to xnano.core.nodes
AbstractNode/NodeKind are shared contracts imported by webui and core
controllers; hosting them inside xnano.tui under a temp module name
violated the layer boundary. Moved verbatim, imports updated, temp
module deleted.
