+++ xnano/webui/nodes.py
@@ -14,7 +14,7 @@ from typing import ClassVar
 
 from xnano._types import Alignment, CharacterModifier, Direction
 from xnano.color import Color, ColorLike
-from xnano.tui._node_base_tmp import AbstractNode, NodeKind
+from xnano.core.nodes import AbstractNode, NodeKind
 
 
 def build_style_attrs(
